fix interface: long as input lengths and output idx dtype; use int throughout inner ds

+++ csrc/grid/counting_sort_cpu.cpp
@@ -11,7 +11,7 @@ void CountingSortCPU(
     at::Tensor sorted_point_idx) {
 
   auto points_a = points.accessor<float, 3>();
-  auto lengths_a = lengths.accessor<int, 1>();
+  auto lengths_a = lengths.accessor<long, 1>();
   auto grid_cell_a = grid_cell.accessor<int, 2>();
   auto grid_idx_a = grid_idx.accessor<int, 2>();
   auto grid_off_a = grid_off.accessor<int, 2>();
